+++ src/app.rs
@@ -1,8 +1,11 @@
 //! Document the command line interface
+use shlex::Shlex;
 use std::collections::HashMap;
+use std::convert::TryFrom;
 use std::env;
 use std::fs::read_to_string;
 use std::path::{Path, PathBuf};
+use std::time::Instant;
 use structopt::StructOpt;
 use two_timer;
 
@@ -10,9 +13,12 @@ use lazy_static::lazy_static;
 use regex::Regex;
 
 use crate::commands::{accounts, balance, commodities, payees, prices, register, statistics};
+use crate::models::Ledger;
 use crate::Error;
 use chrono::NaiveDate;
 use colored::Colorize;
+
+const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 
 #[derive(Debug, StructOpt)]
 enum Command {
@@ -57,6 +63,12 @@ struct Opt {
     #[structopt(subcommand)]
     cmd: Command,
 }
+
+#[derive(Debug, StructOpt)]
+struct Repl {
+    #[structopt(flatten)]
+    options: CommonOpts,
+}
 /// Command line options
 #[derive(Debug, StructOpt, Clone)]
 pub struct CommonOpts {
@@ -96,7 +108,7 @@ pub struct CommonOpts {
     pub no_balance_check: bool,
 
     /// Display the report in the selected currency
-    #[structopt(short = "-X")]
+    #[structopt(long = "--exchange", short = "-X")]
     pub exchange: Option<String>,
 
     /// TODO Date format
@@ -127,6 +139,10 @@ pub struct CommonOpts {
     /// TODO Unrealized losses
     #[structopt(long = "--unrealized-losses")]
     unrealized_losses: Option<String>,
+
+    /// Whether to collapse postings from the same account in the same transaction
+    #[structopt(long = "--collapse")]
+    pub collapse: bool,
 }
 
 impl CommonOpts {
@@ -152,6 +168,7 @@ impl CommonOpts {
             pedantic: false,
             unrealized_gains: None,
             unrealized_losses: None,
+            collapse: false,
         }
     }
 }
@@ -191,9 +208,9 @@ fn init_paths(args: Vec<String>) -> Result<Vec<String>, ()> {
 }
 
 /// Entry point for the command line app
-pub fn run_app(mut args: Vec<String>) -> Result<(), ()> {
+pub fn run_app(input_args: Vec<String>) -> Result<(), ()> {
     let mut config_file = None;
-    let possible_paths = init_paths(args.clone())?;
+    let possible_paths = init_paths(input_args.clone())?;
     for path in possible_paths.iter() {
         let file = Path::new(path);
         if file.exists() {
@@ -201,45 +218,131 @@ pub fn run_app(mut args: Vec<String>) -> Result<(), ()> {
             break;
         }
     }
-    if let Some(file) = config_file {
-        let mut aliases = HashMap::new();
-        aliases.insert("-f".to_string(), "--file".to_string());
-        let contents = read_to_string(file).unwrap();
-        for line in contents.lines() {
-            let option = line.trim_start();
-            match option.chars().nth(0) {
-                Some(c) => match c {
-                    '-' => {
-                        let message = format!("Bad config file {:?}\n{}", file, line);
-                        assert!(line.starts_with("--"), "{}", message);
-                        let mut iter = line.split_whitespace();
-                        let option = iter.next().unwrap();
-                        if !args.iter().any(|x| {
-                            (x == option) | (aliases.get(x).unwrap_or(&String::new()) == option)
-                        }) {
-                            args.push(option.to_string());
-                            let mut rest = String::new();
-                            for arg in iter {
-                                rest.push_str(" ");
-                                rest.push_str(arg);
-                            }
-                            if rest.len() > 0 {
-                                args.push(rest.trim().to_string());
-                            }
+    let args = if let Some(file) = config_file {
+        parse_config_file(file, &input_args)
+    } else {
+        input_args
+    };
+
+    match Opt::from_iter_safe(args.iter()) {
+        Err(error) => match Repl::from_iter_safe(args.iter()) {
+            Ok(opt) => {
+                if opt.options.query.len() > 0 {
+                    error.exit()
+                } else {
+                    println!("dinero-rs v{}", VERSION);
+
+                    let start = Instant::now();
+                    let mut ledger = Ledger::try_from(&opt.options).unwrap();
+                    let duration = start.elapsed();
+                    println!(
+                        "Loaded ledger from {:?} in {:?}",
+                        &opt.options.input_file, duration
+                    );
+
+                    // Start the REPL
+                    let mut rl = rustyline::Editor::<()>::new();
+                    loop {
+                        let readline = rl.readline(">> ");
+                        match readline {
+                            Ok(line) => match line.as_str() {
+                                "exit" | "quit" => break,
+                                "reload" => {
+                                    let start = Instant::now();
+                                    let journal = Ledger::try_from(&opt.options);
+                                    let duration = start.elapsed();
+                                    match journal {
+                                        Ok(j) => {
+                                            println!(
+                                                "Loaded journal from {:?} in {:?}",
+                                                &opt.options.input_file, duration
+                                            );
+                                            ledger = j;
+                                        }
+                                        Err(x) => {
+                                            eprintln!("Journal could not be reloaded. Please check the errors and try again.");
+                                            eprintln!("{}", x);
+                                        }
+                                    }
+                                }
+                                line => match line.trim().is_empty() {
+                                    true => (),
+                                    false => {
+                                        let mut arguments: Vec<String> = Shlex::new(line).collect();
+                                        if !line.starts_with("dinero ") {
+                                            arguments.insert(0, String::from(""))
+                                        }
+                                        let args = if let Some(file) = config_file {
+                                            parse_config_file(file, &arguments)
+                                        } else {
+                                            arguments
+                                        };
+                                        match Opt::from_iter_safe(args) {
+                                            Ok(opt) => {
+                                                match execute_command(opt, Some(ledger.clone())) {
+                                                    Ok(_) => (),
+                                                    Err(x) => eprintln!("{}\nThe above command resulted in an error. {:?}", line,x)
+                                                }
+                                            }
+                                            Err(error) => {
+                                                eprintln!("{}", error);
+                                            }
+                                        }
+                                    }
+                                },
+                            },
+                            Err(_) => break,
                         }
                     }
-                    ';' | '#' | '!' | '%' => (), // a comment
-
-                    _ => panic!("Bad config file {:?}\n{}", file, line),
-                },
-                None => (),
+                }
+                Ok(())
             }
+            Err(_) => error.exit(),
+        },
+        Ok(opt) => execute_command(opt, None),
+    }
+}
+
+fn parse_config_file(file: &Path, input_args: &Vec<String>) -> Vec<String> {
+    let mut args = input_args.clone();
+
+    let mut aliases = HashMap::new();
+    aliases.insert("-f".to_string(), "--file".to_string());
+    let contents = read_to_string(file).unwrap();
+    for line in contents.lines() {
+        let option = line.trim_start();
+        match option.chars().nth(0) {
+            Some(c) => match c {
+                '-' => {
+                    let message = format!("Bad config file {:?}\n{}", file, line);
+                    assert!(line.starts_with("--"), "{}", message);
+                    let mut iter = line.split_whitespace();
+                    let option = iter.next().unwrap();
+                    if !args.iter().any(|x| {
+                        (x == option) | (aliases.get(x).unwrap_or(&String::new()) == option)
+                    }) {
+                        args.push(option.to_string());
+                        let mut rest = String::new();
+                        for arg in iter {
+                            rest.push_str(" ");
+                            rest.push_str(arg);
+                        }
+                        if rest.len() > 0 {
+                            args.push(rest.trim().to_string());
+                        }
+                    }
+                }
+                ';' | '#' | '!' | '%' => (), // a comment
+
+                _ => panic!("Bad config file {:?}\n{}", file, line),
+            },
+            None => (),
         }
     }
-    let opt: Opt = Opt::from_iter(args.iter());
-
+    args
+}
+fn execute_command(opt: Opt, maybe_ledger: Option<Ledger>) -> Result<(), ()> {
     // Print options
-    // println!("{:?}", opt.cmd);
     if let Err(e) = match opt.cmd {
         Command::Balance {
             options,
@@ -249,41 +352,41 @@ pub fn run_app(mut args: Vec<String>) -> Result<(), ()> {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
-            balance::execute(&options, flat, !no_total)
+            balance::execute(&options, maybe_ledger, flat, !no_total)
         }
         Command::Register(options) => {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
-            register::execute(&options)
+            register::execute(&options, maybe_ledger)
         }
         Command::Commodities(options) => {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
 
-            commodities::execute(options.input_file.clone(), &options)
+            commodities::execute(&options, maybe_ledger)
         }
         Command::Payees(options) => {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
 
-            payees::execute(options.input_file.clone(), &options)
+            payees::execute(&options, maybe_ledger)
         }
-        Command::Prices(options) => prices::execute(options.input_file.clone(), &options),
+        Command::Prices(options) => prices::execute(&options, maybe_ledger),
         Command::Accounts(options) => {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
 
-            accounts::execute(options.input_file.clone(), &options)
+            accounts::execute(&options, maybe_ledger)
         }
         Command::Statistics(options) => {
             if options.force_color {
                 env::set_var("CLICOLOR_FORCE", "1");
             }
-            statistics::execute(options.input_file.clone(), &options)
+            statistics::execute(&options, maybe_ledger)
         }
     } {
         let err_str = format!("{}", e);