+++ CHANGELOG.md
@@ -1,10 +1,36 @@
 # Changelog
 Changelog file for dinero-rs project, a command line application for managing finances.
 
-## [0.28.0] - xxx
+## [0.31.0] - xxx
 ### Fixed
 - [Currencies are shown consistently in a report](https://github.com/frosklis/dinero-rs/issues/103)
 
+## [0.30.0] - 2021-08-18
+## Added
+- Show more info when loading the repl
+- Ability to [reload the journal](https://github.com/frosklis/dinero-rs/issues/116) 
+## Fixed
+- [```Some payees were None```](https://github.com/frosklis/dinero-rs/issues/121)
+
+## [0.29.1] - 2021-08-17
+## Changed
+- small improvements on REPL interface
+- improved test coverage
+## [0.29.0] - 2021-08-16
+### Added
+- ```exchange``` option (```-X```) for register reports
+- REPL interface, which is faster than the CLI once everything's loaded
+### Changed
+- Some internal tests now use the ```--init-file``` flag to make sure the environment is properly replicated.
+- Updated dependency from ```assert_cmd to 2.0```
+
+## [0.28.1] - 2021-08-10
+### Fixed
+- The previous crate was created badly.
+
+## [0.28.0] - 2021-08-09
+### Added
+- ```--collapse``` flag to collapse postings with the same currency and account
 ## [0.27.0] - 2021-08-04
 ### Fixed
 - Negative quantities starting with zero now show the negative sign.